move is function tests
